Refactor parser and compiler
