+++ src/compiler.rs
@@ -1,6 +1,10 @@
 use miette::{LabeledSpan, Result, SourceSpan};
 
-use crate::{chunk::Chunk, types::function::Function};
+use crate::{
+    chunk::Chunk,
+    op::Op,
+    types::{function::Function, obj_ref::ObjRef, value::Value},
+};
 #[derive(PartialEq, Debug)]
 struct Local<'a> {
     name: &'a str,
@@ -23,6 +27,11 @@ pub struct Compiler<'a> {
 pub struct ResolveResult {
     pub slot: usize,
     pub initialized: bool,
+}
+pub struct Jump {
+    op: fn(u16) -> Op,
+    location: SourceSpan,
+    position: usize,
 }
 
 impl<'a> Compiler<'a> {
@@ -47,19 +56,17 @@ impl<'a> Compiler<'a> {
         self.scope_depth += 1;
     }
 
-    pub fn end_scope(&mut self) -> usize {
+    pub fn end_scope(&mut self, location: SourceSpan) {
         self.scope_depth -= 1;
 
-        let mut popped: usize = 0;
         while let Some(last) = self.locals.last() {
             if last.depth.is_none_or(|s| s > self.scope_depth) {
                 self.locals.pop();
-                popped += 1;
+                self.chunk.write(Op::Pop, location);
             } else {
                 break;
             }
         }
-        popped
     }
 
     pub fn add_local(&mut self, name: &'a str, location: SourceSpan) -> Result<()> {
@@ -75,8 +82,11 @@ impl<'a> Compiler<'a> {
     }
 
     pub fn mark_latest_initialized(&mut self) {
-        if let Some(last) = self.locals.last_mut() {
-            last.depth = Some(self.scope_depth);
+        if self.scope_depth > 0 {
+            // happens in global function declaration
+            if let Some(last) = self.locals.last_mut() {
+                last.depth = Some(self.scope_depth);
+            }
         }
     }
 
@@ -102,6 +112,72 @@ impl<'a> Compiler<'a> {
 
     pub fn end_compiler(self) -> Function {
         Function::new(0, self.chunk, None) // TODO: Real names for real functions
+    }
+
+    pub fn define_variable(&mut self, global_idx: Option<u8>, location: SourceSpan) {
+        if let Some(const_idx) = global_idx {
+            self.chunk.write(Op::DefineGlobal(const_idx), location);
+        } else {
+            self.mark_latest_initialized();
+        }
+    }
+
+    pub fn declare_variable(&mut self, name: &'a str, location: SourceSpan) -> Result<()> {
+        if self.is_local() {
+            if self.has_variable_in_current_scope(name) {
+                miette::bail!(
+                    labels = vec![LabeledSpan::at(location, "here")],
+                    "Already a variable with this name in this scope"
+                )
+            }
+            self.add_local(name, location)?;
+        }
+        Ok(())
+    }
+
+    pub fn identifier_constant(&mut self, name: ObjRef) -> u8 {
+        self.chunk.add_constant(Value::Obj(name))
+    }
+
+    pub fn emit_constant(&mut self, value: Value, location: SourceSpan) {
+        let idx = self.chunk.add_constant(value);
+        self.chunk.write(Op::Constant(idx), location);
+    }
+
+    pub fn emit_jump(&mut self, op: fn(u16) -> Op, location: SourceSpan) -> Jump {
+        let position = self.chunk.code.len();
+        self.chunk.write(op(0), location);
+        Jump {
+            op,
+            location,
+            position,
+        }
+    }
+
+    pub fn emit_loop(&mut self, loop_start: usize, location: SourceSpan) -> Result<()> {
+        let jump_length = self.chunk.code.len() - loop_start;
+        if let Ok(jump_length) = u16::try_from(jump_length) {
+            self.chunk.write(Op::Loop(jump_length), location);
+            Ok(())
+        } else {
+            miette::bail!(
+                labels = vec![LabeledSpan::at(location, "here")],
+                "Loop body too large."
+            )
+        }
+    }
+
+    pub fn patch_jump(&mut self, jump: Jump) -> Result<()> {
+        let jump_length = self.chunk.code.len() - jump.position;
+        if let Ok(jump_length) = u16::try_from(jump_length) {
+            self.chunk.code[jump.position] = (jump.op)(jump_length);
+            Ok(())
+        } else {
+            miette::bail!(
+                labels = vec![LabeledSpan::at(jump.location, "here")],
+                "Too much code to jump over"
+            )
+        }
     }
 }
 
@@ -193,7 +269,8 @@ mod tests {
     }
 
     #[test]
-    fn end_scope_returns_correct_count() {
+    fn end_scope_writes_enough_pops() {
+        let location = SourceSpan::from((0, 0));
         let mut compiler = Compiler {
             locals: vec![
                 Local {
@@ -213,7 +290,10 @@ mod tests {
             function_type: FunctionType::Script,
             chunk: Chunk::new(),
         };
-        assert_eq!(compiler.end_scope(), 2);
+        compiler.end_scope(location);
+        assert_eq!(compiler.chunk.code.len(), 2);
+        assert_eq!(compiler.chunk.code[0], Op::Pop);
+        assert_eq!(compiler.chunk.code[1], Op::Pop);
     }
 
     #[test]
